requirements.txt
Django==3.0.6
django-allauth==0.41.0
django-cities-light==3.6.0
django-cors-headers==3.3.0
django-countries==6.1.2
django-crispy-forms==1.9.1
django-phonenumber-field==4.0.0
django-rest-framework==0.1.0
djangorestframework==3.11.0
misaka==2.1.1
phonenumbers==8.12.4
Pillow==7.0.0
django-notifications==0.1.dev0
